refactor(provider): keep history index helpers local
