+++ crates/storage/storage-api/src/history.rs
@@ -1,7 +1,7 @@
 use alloy_primitives::{Address, BlockNumber, B256};
 use auto_impl::auto_impl;
 use core::ops::{RangeBounds, RangeInclusive};
-use reth_db_api::{cursor::DbCursorRO, models::BlockNumberAddress, table::Table, BlockNumberList};
+use reth_db_api::models::BlockNumberAddress;
 use reth_db_models::AccountBeforeTx;
 use reth_primitives_traits::StorageEntry;
 use reth_storage_errors::provider::ProviderResult;
@@ -59,57 +59,6 @@ pub trait HistoryReader: Send {
         block_number: BlockNumber,
         lowest_available_block_number: Option<BlockNumber>,
     ) -> ProviderResult<HistoryInfo>;
-}
-
-/// Computes the position of the first history entry at or after `block_number`.
-#[inline]
-pub fn compute_history_rank(
-    chunk: &BlockNumberList,
-    block_number: BlockNumber,
-) -> (u64, Option<u64>) {
-    let mut rank = chunk.rank(block_number);
-    if rank.checked_sub(1).and_then(|rank| chunk.select(rank)) == Some(block_number) {
-        rank -= 1;
-    }
-    (rank, chunk.select(rank))
-}
-
-/// Returns whether a history lookup needs to inspect the preceding shard.
-#[inline]
-pub fn needs_prev_shard_check(
-    rank: u64,
-    found_block: Option<BlockNumber>,
-    block_number: BlockNumber,
-) -> bool {
-    rank == 0 && found_block != Some(block_number)
-}
-
-/// Looks up a key in a sharded history table.
-pub fn history_info<T, K, C>(
-    cursor: &mut C,
-    key: K,
-    block_number: BlockNumber,
-    key_filter: impl Fn(&K) -> bool,
-    lowest_available_block_number: Option<BlockNumber>,
-) -> ProviderResult<HistoryInfo>
-where
-    T: Table<Key = K, Value = BlockNumberList>,
-    C: DbCursorRO<T>,
-{
-    if let Some(chunk) = cursor.seek(key)?.filter(|(key, _)| key_filter(key)).map(|entry| entry.1) {
-        let (rank, found_block) = compute_history_rank(&chunk, block_number);
-        let is_before_first_write = needs_prev_shard_check(rank, found_block, block_number) &&
-            !cursor.prev()?.is_some_and(|(key, _)| key_filter(&key));
-        Ok(HistoryInfo::from_lookup(
-            found_block,
-            is_before_first_write,
-            lowest_available_block_number,
-        ))
-    } else if lowest_available_block_number.is_some() {
-        Ok(HistoryInfo::MaybeInPlainState)
-    } else {
-        Ok(HistoryInfo::NotYetWritten)
-    }
 }
 
 /// History Writer
